Backend-Diseases-Priority/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==23.0.0
torch==2.3.0
tensorflow==2.15.0


scikit-learn==1.4.2
joblib==1.4.2
numpy>=1.23.5,<2.0.0
pandas==1.5.3
matplotlib==3.9.3
seaborn==0.13.2
scipy==1.14.1
opencv-python==4.8.1.78

transformers==4.41.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
sacremoses==0.1.1

python-dotenv==0.19.2
requests==2.31.0

python-multipart